move exception into separate file
